[FEAT] #169 - BaseActivity, BaseFragment 구현 및 적용
