+++ app/src/main/java/co/kr/bemyplan/ui/base/BaseActivity.kt
@@ -1,0 +1,19 @@
+package co.kr.bemyplan.ui.base
+
+import android.os.Bundle
+import android.util.Log
+import androidx.annotation.LayoutRes
+import androidx.appcompat.app.AppCompatActivity
+import androidx.databinding.DataBindingUtil
+import androidx.databinding.ViewDataBinding
+
+abstract class BaseActivity<T : ViewDataBinding>(@LayoutRes val layoutRes: Int) :
+    AppCompatActivity() {
+    lateinit var binding: T // Data Binding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = DataBindingUtil.setContentView(this, layoutRes)
+        binding.lifecycleOwner = this
+    }
+}
